refactor(server): migrate openai responses to chain dispatch (#619)
## Summary
After code review identified issues with the chain dispatch refactor,
this PR completes the migration of OpenAI Responses API handlers to use
the unified dispatch pattern, eliminating 360+ lines of duplicate
protocol conversion code.

### Major

- **Migrated Responses API to chain dispatch**: `openai_responses.go`
now delegates to `protocol_dispatch.go` handlers instead of implementing
direct request forwarding
- **Added streaming/non-streaming handlers**: New
`streamOpenAIResponses()` and `nonstreamOpenAIResponses()` in
`protocol_dispatch.go` handle Responses API passthrough
- **Fixed model tracking bug**: Changed
`recorder.RecordResponse(provider, actualModel)` to
`recorder.RecordResponse(provider, reqCtx.RequestModel)` for consistent
tracking

### Architecture Changes

**Before** (`openai_responses.go`):
```go
// Direct dispatch with duplicated forwarding logic
switch target {
case protocol.TypeOpenAIResponses:
s.handleResponsesStreamingRequest(...)
case protocol.TypeOpenAIChat:
// Inline Chat→Responses conversion (100+ lines)
case protocol.TypeAnthropicV1:
// Inline Anthropic→Responses conversion (80+ lines)
...
}

After (protocol_dispatch.go):
// Chain dispatch via unified handlers
func (s *Server) dispatchChainFromResponses(...) {
switch target {
case protocol.TypeAnthropicV1:
s.streamAnthropicV1ToResponses(...)
case protocol.TypeOpenAIChat:
s.dispatchOpenAIChatFromResponses(...)
case protocol.TypeOpenAIResponses:
s.streamOpenAIResponses(...) // New passthrough handler
}
}
```

### Minor

- Code cleanup: removed unused imports from openai_responses.go
- Chore commit: minor formatting/organization
