+++ internal/server/openai_responses.go
@@ -1,21 +1,15 @@
 package server
 
 import (
-	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
 	"net/http"
 
 	"github.com/anthropics/anthropic-sdk-go"
 	"github.com/gin-gonic/gin"
 	"github.com/openai/openai-go/v3"
 	"github.com/openai/openai-go/v3/packages/param"
-	"github.com/openai/openai-go/v3/packages/ssestream"
 	"github.com/openai/openai-go/v3/responses"
-	"github.com/sirupsen/logrus"
-	. "github.com/tingly-dev/tingly-box/internal/protocol/stream"
 
 	"github.com/tingly-dev/tingly-box/internal/loadbalance"
 	"github.com/tingly-dev/tingly-box/internal/protocol"
@@ -136,10 +130,10 @@ func (s *Server) HandleResponsesCreate(c *gin.Context) {
 	req.ResponseNewParams = params
 	// req.Model is replaced with actualModel (resolved backend model) from this point on
 	req.Model = actualModel
-	s.ResponsesCreate(c, scenarioType, provider, req, rule.RequestModel, maxAllowed)
+	s.ResponsesCreate(c, scenarioType, provider, rule, req, rule.RequestModel, maxAllowed)
 }
 
-func (s *Server) ResponsesCreate(c *gin.Context, scenarioType typ.RuleScenario, provider *typ.Provider, req protocol.ResponseCreateRequest, responseModel string, maxAllowed int) {
+func (s *Server) ResponsesCreate(c *gin.Context, scenarioType typ.RuleScenario, provider *typ.Provider, rule *typ.Rule, req protocol.ResponseCreateRequest, responseModel string, maxAllowed int) {
 	actualModel := req.Model
 	isStreaming := req.Stream
 
@@ -186,102 +180,11 @@ func (s *Server) ResponsesCreate(c *gin.Context, scenarioType typ.RuleScenario, 
 		return
 	}
 
-	// Dispatch based on target
-	switch target {
-	case protocol.TypeOpenAIResponses:
-		params := reqCtx.Request.(*responses.ResponseNewParams)
-		if isStreaming {
-			s.handleResponsesStreamingRequest(c, provider, *params, responseModel, actualModel)
-		} else {
-			s.handleResponsesNonStreamingRequest(c, provider, *params, responseModel, actualModel)
-		}
-
-	case protocol.TypeOpenAIChat:
-		chatReq := reqCtx.Request.(*openai.ChatCompletionNewParams)
-		if isStreaming {
-			wrapper := s.clientPool.GetOpenAIClient(provider, string(chatReq.Model))
-			fc := NewForwardContext(c.Request.Context(), provider)
-			chatStream, cancel, err := ForwardOpenAIChatStream(fc, wrapper, chatReq)
-			if cancel != nil {
-				defer cancel()
-			}
-			if err != nil {
-				s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(0, 0, 0), err)
-				c.JSON(http.StatusInternalServerError, ErrorResponse{
-					Error: ErrorDetail{Message: "Failed to create streaming request: " + err.Error(), Type: "api_error"},
-				})
-				return
-			}
-			usage, err := HandleOpenAIChatToResponsesStream(c, chatStream, responseModel)
-			s.trackUsageWithTokenUsage(c, usage, err)
-		} else {
-			wrapper := s.clientPool.GetOpenAIClient(provider, string(chatReq.Model))
-			fc := NewForwardContext(nil, provider)
-			chatResp, _, err := ForwardOpenAIChat(fc, wrapper, chatReq)
-			if err != nil {
-				s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(0, 0, 0), err)
-				c.JSON(http.StatusInternalServerError, ErrorResponse{
-					Error: ErrorDetail{Message: "Failed to forward request: " + err.Error(), Type: "api_error"},
-				})
-				return
-			}
-			inputTokens := chatResp.Usage.PromptTokens
-			outputTokens := chatResp.Usage.CompletionTokens
-			cacheTokens := chatResp.Usage.PromptTokensDetails.CachedTokens
-			s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(int(inputTokens), int(outputTokens), int(cacheTokens)), nil)
-			c.JSON(http.StatusOK, buildResponsesPayloadFromChat(chatResp, responseModel, actualModel))
-		}
-
-	case protocol.TypeAnthropicV1:
-		anthropicReq := reqCtx.Request.(*anthropic.MessageNewParams)
-		if isStreaming {
-			wrapper := s.clientPool.GetAnthropicClient(provider, string(anthropicReq.Model))
-			fc := NewForwardContext(c.Request.Context(), provider)
-			anthropicStream, cancel, err := ForwardAnthropicV1Stream(fc, wrapper, anthropicReq)
-			if cancel != nil {
-				defer cancel()
-			}
-			if err != nil {
-				s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(0, 0, 0), err)
-				c.JSON(http.StatusInternalServerError, ErrorResponse{
-					Error: ErrorDetail{Message: "Failed to create streaming request: " + err.Error(), Type: "api_error"},
-				})
-				return
-			}
-
-			hc := protocol.NewHandleContext(c, responseModel)
-			firstTokenRecorded := false
-			hc.WithOnStreamEvent(func(_ interface{}) error {
-				if !firstTokenRecorded {
-					SetFirstTokenTime(c)
-					firstTokenRecorded = true
-				}
-				return nil
-			})
-			usage, err := HandleAnthropicToOpenAIResponsesStream(hc, anthropicStream, responseModel)
-			s.trackUsageWithTokenUsage(c, usage, err)
-		} else {
-			wrapper := s.clientPool.GetAnthropicClient(provider, string(anthropicReq.Model))
-			fc := NewForwardContext(nil, provider)
-			anthropicResp, cancel, err := ForwardAnthropicV1(fc, wrapper, anthropicReq)
-			if cancel != nil {
-				defer cancel()
-			}
-			if err != nil {
-				s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(0, 0, 0), err)
-				c.JSON(http.StatusInternalServerError, ErrorResponse{
-					Error: ErrorDetail{Message: "Failed to forward request: " + err.Error(), Type: "api_error"},
-				})
-				return
-			}
-
-			cacheTokens := int(anthropicResp.Usage.CacheReadInputTokens)
-			s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(int(anthropicResp.Usage.InputTokens), int(anthropicResp.Usage.OutputTokens), cacheTokens), nil)
-			c.JSON(http.StatusOK, buildResponsesPayloadFromAnthropic(anthropicResp, responseModel, actualModel))
-		}
-	}
+	// Use unified dispatch
+	s.dispatchChainFromResponses(c, reqCtx, rule, provider, isStreaming, nil)
 }
 
+// buildResponsesPayloadFromChat converts a Chat completion response to Responses API format
 func buildResponsesPayloadFromChat(resp *openai.ChatCompletion, responseModel, actualModel string) map[string]any {
 	model := responseModel
 	if model == "" {
@@ -316,6 +219,7 @@ func buildResponsesPayloadFromChat(resp *openai.ChatCompletion, responseModel, a
 	}
 }
 
+// buildResponsesPayloadFromAnthropic converts an Anthropic message response to Responses API format
 func buildResponsesPayloadFromAnthropic(resp *anthropic.Message, responseModel, actualModel string) map[string]any {
 	model := responseModel
 	if model == "" {
@@ -366,266 +270,6 @@ func buildResponsesPayloadFromAnthropic(resp *anthropic.Message, responseModel, 
 			"total_tokens":  resp.Usage.InputTokens + resp.Usage.OutputTokens,
 		},
 	}
-}
-
-// handleResponsesNonStreamingRequest handles non-streaming Responses API requests
-func (s *Server) handleResponsesNonStreamingRequest(c *gin.Context, provider *typ.Provider, params responses.ResponseNewParams, responseModel, actualModel string) {
-	// Forward request to provider
-	var response *responses.Response
-	var err error
-	var cancel context.CancelFunc
-
-	wrapper := s.clientPool.GetOpenAIClient(provider, string(params.Model))
-	fc := NewForwardContext(nil, provider)
-	response, cancel, err = ForwardOpenAIResponses(fc, wrapper, params)
-	if cancel != nil {
-		defer cancel()
-	}
-
-	if err != nil {
-		// Track error with no usage
-		s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(0, 0, 0), err)
-		c.JSON(http.StatusInternalServerError, ErrorResponse{
-			Error: ErrorDetail{
-				Message: "Failed to forward request: " + err.Error(),
-				Type:    "api_error",
-			},
-		})
-		return
-	}
-
-	// Extract usage from response
-	inputTokens := int64(response.Usage.InputTokens)
-	outputTokens := int64(response.Usage.OutputTokens)
-	cacheTokens := int64(response.Usage.InputTokensDetails.CachedTokens)
-
-	// Track usage
-	s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(int(inputTokens), int(outputTokens), int(cacheTokens)), nil)
-
-	// Override model in response if needed
-	if responseModel != actualModel {
-		// Create a copy of the response with updated model
-		responseJSON, _ := json.Marshal(response)
-		var responseMap map[string]any
-		if err := json.Unmarshal(responseJSON, &responseMap); err == nil {
-			responseMap["model"] = responseModel
-			c.JSON(http.StatusOK, responseMap)
-			return
-		}
-	}
-
-	// Return response as-is
-	c.JSON(http.StatusOK, response)
-}
-
-// handleResponsesStreamingRequest handles streaming Responses API requests
-func (s *Server) handleResponsesStreamingRequest(c *gin.Context, provider *typ.Provider, params responses.ResponseNewParams, responseModel, actualModel string) {
-	// Create streaming request with request context for proper cancellation
-	wrapper := s.clientPool.GetOpenAIClient(provider, params.Model)
-	fc := NewForwardContext(c.Request.Context(), provider)
-	stream, cancel, err := ForwardOpenAIResponsesStream(fc, wrapper, params)
-	if cancel != nil {
-		defer cancel()
-	}
-	if err != nil {
-		// Track error with no usage
-		s.trackUsageFromContext(c, 0, 0, err)
-		c.JSON(http.StatusInternalServerError, ErrorResponse{
-			Error: ErrorDetail{
-				Message: "Failed to create streaming request: " + err.Error(),
-				Type:    "api_error",
-			},
-		})
-		return
-	}
-
-	// Handle the streaming response
-	hc := protocol.NewHandleContext(c, responseModel)
-	firstTokenRecorded := false
-	hc.WithOnStreamEvent(func(_ interface{}) error {
-		if !firstTokenRecorded {
-			SetFirstTokenTime(c)
-			firstTokenRecorded = true
-		}
-		return nil
-	})
-	usage, err := HandleOpenAIResponsesStream(hc, stream, responseModel)
-
-	// Track usage from stream handler
-	s.trackUsageWithTokenUsage(c, usage, err)
-}
-
-// handleResponsesStreamResponse processes the streaming response and sends it to the client
-func (s *Server) handleResponsesStreamResponse(c *gin.Context, stream *ssestream.Stream[responses.ResponseStreamEventUnion], responseModel, actualModel string) {
-	// Accumulate usage from stream chunks
-	var inputTokens, outputTokens int64
-	var hasUsage bool
-
-	defer func() {
-		if r := recover(); r != nil {
-			logrus.Errorf("Panic in streaming handler: %v", r)
-			if hasUsage {
-				s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(int(inputTokens), int(outputTokens), 0), fmt.Errorf("panic: %v", r))
-			}
-			if c.Writer != nil {
-				c.Writer.WriteHeader(http.StatusInternalServerError)
-				c.Writer.Write([]byte("data: {\"error\":{\"message\":\"Internal streaming error\",\"type\":\"internal_error\"}}\n\n"))
-				if flusher, ok := c.Writer.(http.Flusher); ok {
-					flusher.Flush()
-				}
-			}
-		}
-		if stream != nil {
-			if err := stream.Close(); err != nil {
-				logrus.Errorf("Error closing stream: %v", err)
-			}
-		}
-	}()
-
-	// Set SSE headers
-	c.Header("Content-Type", "text/event-stream")
-	c.Header("Cache-Control", "no-cache")
-	c.Header("Connection", "keep-alive")
-	c.Header("Access-Control-Allow-Origin", "*")
-	c.Header("Access-Control-Allow-Headers", "Cache-Control")
-
-	flusher, ok := c.Writer.(http.Flusher)
-	if !ok {
-		c.JSON(http.StatusInternalServerError, ErrorResponse{
-			Error: ErrorDetail{
-				Message: "Streaming not supported by this connection",
-				Type:    "api_error",
-				Code:    "streaming_unsupported",
-			},
-		})
-		return
-	}
-
-	// Process the stream with context cancellation checking
-	c.Stream(func(w io.Writer) bool {
-		// Check context cancellation first
-		select {
-		case <-c.Request.Context().Done():
-			logrus.Debug("Client disconnected, stopping Responses stream")
-			return false
-		default:
-		}
-
-		// Try to get next event
-		if !stream.Next() {
-			// Stream ended
-			return false
-		}
-
-		event := stream.Current()
-
-		// Accumulate usage from completed events
-		if event.Response.Usage.InputTokens > 0 {
-			inputTokens = event.Response.Usage.InputTokens
-			hasUsage = true
-		}
-		if event.Response.Usage.OutputTokens > 0 {
-			outputTokens = event.Response.Usage.OutputTokens
-		}
-
-		// Use the event type as the SSE event type (e.g., "response.created", "response.output_text.delta")
-		SSEventOpenAI(c, event.Type, event, responseModel)
-		flusher.Flush()
-		return true
-	})
-
-	// Check for stream errors
-	if err := stream.Err(); err != nil {
-		// Check if it was a client cancellation
-		if errors.Is(err, context.Canceled) {
-			logrus.Debug("Responses stream canceled by client")
-			if hasUsage {
-				s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(int(inputTokens), int(outputTokens), 0), err)
-			}
-			return
-		}
-
-		logrus.Errorf("Stream error: %v", err)
-		if hasUsage {
-			s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(int(inputTokens), int(outputTokens), 0), err)
-		}
-
-		errorChunk := map[string]any{
-			"error": map[string]any{
-				"message": err.Error(),
-				"type":    "stream_error",
-				"code":    "stream_failed",
-			},
-		}
-
-		errorJSON, marshalErr := json.Marshal(errorChunk)
-		if marshalErr != nil {
-			logrus.Errorf("Failed to marshal error chunk: %v", marshalErr)
-			c.Writer.Write([]byte("data: {\"error\":{\"message\":\"Failed to marshal error\",\"type\":\"internal_error\"}}\n\n"))
-		} else {
-			c.Writer.Write([]byte(fmt.Sprintf("data: %s\n\n", string(errorJSON))))
-		}
-		flusher.Flush()
-		return
-	}
-
-	// Track successful streaming completion
-	if hasUsage {
-		s.trackUsageWithTokenUsage(c, protocol.NewTokenUsageWithCache(int(inputTokens), int(outputTokens), 0), nil)
-	}
-
-	// Send the final [DONE] message
-	c.Writer.Write([]byte("data: [DONE]\n\n"))
-	flusher.Flush()
-}
-
-// SSEventOpenAI sends an SSE event with the given data
-// If the data is a ResponseStreamEventUnion, it uses the raw JSON to avoid
-// serializing empty fields from the union type
-// If modelOverride is provided and the event contains a response object with a model field,
-// it will be overridden in the JSON output
-func SSEventOpenAI(c *gin.Context, t string, data any, modelOverride ...string) error {
-	var jsonBytes []byte
-
-	// For ResponseStreamEventUnion, use RawJSON() to avoid serializing all empty union fields
-	if event, ok := data.(responses.ResponseStreamEventUnion); ok {
-		rawJSON := event.RawJSON()
-		if rawJSON != "" {
-			jsonBytes = []byte(rawJSON)
-			// Apply model override if provided
-			if len(modelOverride) > 0 && modelOverride[0] != "" {
-				var parsed map[string]any
-				if err := json.Unmarshal(jsonBytes, &parsed); err == nil {
-					// Check if this event has a response field with a model
-					if response, ok := parsed["response"].(map[string]any); ok {
-						if model, ok := response["model"].(string); ok && model != "" {
-							response["model"] = modelOverride[0]
-							modified, err := json.Marshal(parsed)
-							if err == nil {
-								jsonBytes = modified
-							}
-						}
-					}
-				}
-			}
-		} else {
-			// Fallback to regular marshaling if RawJSON is empty
-			var err error
-			jsonBytes, err = json.Marshal(event)
-			if err != nil {
-				return err
-			}
-		}
-	} else {
-		var err error
-		jsonBytes, err = json.Marshal(data)
-		if err != nil {
-			return err
-		}
-	}
-
-	c.Writer.WriteString(fmt.Sprintf("data: %s\n\n", jsonBytes))
-	return nil
 }
 
 // convertToResponsesParams converts raw JSON to OpenAI SDK params format
